requirements.txt
pycairo
scipy
pillow
scikit-image